chore(lotto): reorganize Lotto.js folder
- move Lotto.js folder to top-level of src directory
