+++ __tests__/LottoTest.js
@@ -1,4 +1,4 @@
-import Lotto from '../src/models/Lotto.js';
+import Lotto from '../src/Lotto';
 
 describe('로또 클래스 테스트', () => {
   test('로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.', () => {
