+++ src/services/LottoService.js
@@ -1,5 +1,5 @@
 import { Random } from '@woowacourse/mission-utils';
-import Lotto from '../models/Lotto.js';
+import Lotto from '../Lotto.js';
 import {
   LOTTO_MAX_NUMBER,
   LOTTO_MIN_NUMBER,
